Hard fork of missing repo
I generally prefer to use the github fork option where possible to give
credits to the original developer

In this scenario, the original github repo is a dead link and I can't
find any repo that matches the name that has the same code, indicating
that the repo was either deleted, made private, etc.

the github org that made the account also seems to have fallen off the
face of the earth. WMM is expiring in a few short months and this repo
needs new coefficients. It's the most popular wmm repo on crates.io, so
i hope to create an easy upgrade path for when, in 2025, this repo will
start returning Err variants (except for the few clever people who
anticipated this and clamped their input at Jan 31, 2024)

My intention is to keep a valid `wmm2` repo that is compatible with
`wmm`, but just has updated coefficients.